+++ track.h
@@ -4,6 +4,8 @@
 #include <vector>
 #include <string>
 
+#include <stdio.h> // for off_t
+
 class Atom;
 class AVCodecContext;
 class AVCodec;
@@ -11,10 +13,10 @@ class AVCodec;
 class Codec {
 public:
     std::string name;
-    void parse(Atom *trak, std::vector<int> &offsets, Atom *mdat);
-    bool matchSample(unsigned char *start, int maxlength);
-    int getLength(unsigned char *start, int maxlength);
-    bool isKeyframe(unsigned char *start, int maxlength);
+    void parse(Atom *trak, std::vector<off_t> &offsets, Atom *mdat);
+    bool matchSample(unsigned char *start, unsigned int maxlength);
+    int getLength(unsigned char *start, unsigned int maxlength);
+    bool isKeyframe(unsigned char *start, unsigned int maxlength);
     //used by: mp4a
     int mask1;
     int mask0;
@@ -30,7 +32,7 @@ public:
     Codec codec;
 
     std::vector<int> times;
-    std::vector<int> offsets;
+    std::vector<off_t> offsets;
     std::vector<int> sizes;
     std::vector<int> keyframes; //0 based!
 
@@ -43,7 +45,7 @@ public:
     std::vector<int> getSampleTimes(Atom *t);
     std::vector<int> getKeyframes(Atom *t);
     std::vector<int> getSampleSizes(Atom *t);
-    std::vector<int> getChunkOffsets(Atom *t);
+    std::vector<off_t> getChunkOffsets(Atom *t);
     std::vector<int> getSampleToChunk(Atom *t, int nchunks);
 
     void saveSampleTimes();
